requirements.txt
blinker==1.4
docutils==0.12
feedgenerator==1.8
Jinja2==2.8
MarkupSafe==0.23
olefile==0.44
pelican==3.6.3
pelican-readtime==0.1.2
pelican-youtube==0.2.0
Pillow==4.0.0
Pygments==2.1.3
python-dateutil==2.5.3
pytz==2016.4
six==1.10.0
Unidecode==0.4.19
